edited line 1 to test-3.txt to demonstrate descriptions in commit
above, is the concise message describing the reason for the commit.
here, i am describing in more detail what the commit was made for, i.e.
providing a description.

if i needed to put more information in the description, i would create
a new paragraph like this. i am making sure to stay inside the line
limits to make sure the commit and description are readable